refactor: split HtmlMessageText into single-responsibility components
Extract MxcImage to shared/widgets for reuse, HtmlSpanBuilder for DOM
walking, and LinkableSpanBuilder for URL auto-linking and mention pills.
HtmlMessageText becomes a thin caching shell that delegates to the builders.
